+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixes
+
+- Fix dashboard not found errors on delete
+
 ## [0.72.0] - 2026-06-22
 
 ### Changed